+++ History.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Support `deprecate.property(fn, prop, message)`
+
 0.3.0 / 2014-06-16
 ==================